Move Segment KnownTraits.ts to src/types
because it's used by multiple features. and it isn't a string hook.
