+++ flex-template-future-cc-demo/src/feature-library/segment-tab/custom-components/TraitTags.tsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react';
 import { Box, Stack, Badge, SkeletonLoader } from '@twilio-paste/core';
 
 import { SegmentTraits } from '../types/Segment/SegmentTraits';
-import { KnownTraits } from '../flex-hooks/strings/KnownTraits';
+import { KnownTraits } from '../../../types/segment/KnownTraits';
 
 type TraitTagsProps = {
   traits: SegmentTraits;
