+++ flex-template-future-cc-demo/src/types/segment/KnownTraits.ts
